Update docker-compose.yml
Added volume "c:/ProgramData/TempWorks/temp/shared-temp:c:/app/shared-temp" to the api-server volumes. Removed 1809 from the version on api-jobservice-server. It is not needed and will cause the compose to error since you will be specifying the version twice. In the volumes for the "api-jobservice-server" I removed the "3" from "c:/ProgramData/TempWorks/config/api3-jobservice-server:c:/app/config" no that it matches the naming convention of the directory that we have everywhere else for this config.